Correctly renaming some files namespaces and locations
