+++ Jogoteca.Web/Repository/Implementations/GameRepository.cs
@@ -1,7 +1,7 @@
 using Jogoteca.Repository;
 using Jogoteca.Repository.Interfaces;
 using Jogoteca.Models.Entities;
-using Jogoteca.Models.DbContexts;
+using Jogoteca.DbContexts;
 
 namespace Jogoteca.Web.Repository.Implementations
 {
