+++ Jogoteca.Web/Service/Interfaces/IGameService.cs
@@ -1,6 +1,8 @@
+using Jogoteca.Models.Entities;
+
 namespace Jogoteca.Service.Interfaces
 {
-    public interface IGameService
+    public interface IGameService : IGenericService<Game>
     {
          
     }
